Getters/setters now recieve resolve/reject functions instead of working from conditional returns.
Getters will recieve the shadow as the first argument and a resolve function as the second. Resolve will override the shadow return with it's own argument.
Setters will now recieve the shadow as the first argument, the new value as the second, and a reject function as the third. Reject prevents the set from taking place.
Code quality was also improved, shortening some functions, better indenting and variables, and applying some techniques I've learned in the past few months. Also, you can now call GSG without the `new` operator and nothing catches fire.